Move format_version tests out of package_test.py
package_test.py is for CI package integrity checks and must run without
pytest. Move the parametrized format_version tests to their own file.
